Detailed Permissions for Content Libraries

+++ openedx/core/djangoapps/content_libraries/views.py
@@ -1,23 +1,26 @@
 """
 REST API for Blockstore-based content libraries
 """
-
 from functools import wraps
 import logging
 
+from django.contrib.auth import get_user_model
+from django.contrib.auth.models import Group
+from django.shortcuts import get_object_or_404
 from opaque_keys.edx.locator import LibraryLocatorV2, LibraryUsageLocatorV2
 from organizations.models import Organization
 from rest_framework import status
-from rest_framework.exceptions import NotFound, ValidationError
+from rest_framework.exceptions import NotFound, PermissionDenied, ValidationError
 from rest_framework.parsers import MultiPartParser
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from openedx.core.lib.api.view_utils import view_auth_classes
-from . import api
-from .serializers import (
+from openedx.core.djangoapps.content_libraries import api, permissions
+from openedx.core.djangoapps.content_libraries.serializers import (
     ContentLibraryMetadataSerializer,
     ContentLibraryUpdateSerializer,
+    ContentLibraryPermissionLevelSerializer,
+    ContentLibraryPermissionSerializer,
     LibraryXBlockCreationSerializer,
     LibraryXBlockMetadataSerializer,
     LibraryXBlockTypeSerializer,
@@ -25,7 +28,9 @@ from .serializers import (
     LibraryXBlockStaticFileSerializer,
     LibraryXBlockStaticFilesSerializer,
 )
+from openedx.core.lib.api.view_utils import view_auth_classes
 
+User = get_user_model()
 log = logging.getLogger(__name__)
 
 
@@ -62,16 +67,19 @@ class LibraryRootView(APIView):
 
     def get(self, request):
         """
-        Return a list of all content libraries. This is a temporary view for
-        development.
+        Return a list of all content libraries that the user has permission to
+        view. This is a temporary view for development and returns at most 50
+        libraries.
         """
-        result = api.list_libraries()
+        result = api.list_libraries_for_user(request.user)
         return Response(ContentLibraryMetadataSerializer(result, many=True).data)
 
     def post(self, request):
         """
         Create a new content library.
         """
+        if not request.user.has_perm(permissions.CAN_CREATE_CONTENT_LIBRARY):
+            raise PermissionDenied
         serializer = ContentLibraryMetadataSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         data = serializer.validated_data
@@ -103,6 +111,7 @@ class LibraryDetailsView(APIView):
         Get a specific content library
         """
         key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
         result = api.get_library(key)
         return Response(ContentLibraryMetadataSerializer(result).data)
 
@@ -112,6 +121,7 @@ class LibraryDetailsView(APIView):
         Update a content library
         """
         key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY)
         serializer = ContentLibraryUpdateSerializer(data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         api.update_library(key, **serializer.validated_data)
@@ -124,7 +134,94 @@ class LibraryDetailsView(APIView):
         Delete a content library
         """
         key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_DELETE_THIS_CONTENT_LIBRARY)
         api.delete_library(key)
+        return Response({})
+
+
+@view_auth_classes()
+class LibraryTeamView(APIView):
+    """
+    View to get the list of users/groups who can access and edit the content
+    library.
+
+    Note also the 'allow_public_' settings which can be edited by PATCHing the
+    library itself (LibraryDetailsView.patch).
+    """
+    @convert_exceptions
+    def get(self, request, lib_key_str):
+        """
+        Get the list of users and groups who have permissions to view and edit
+        this library.
+        """
+        key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY_TEAM)
+        team = api.get_library_team(key)
+        return Response(ContentLibraryPermissionSerializer(team, many=True).data)
+
+
+@view_auth_classes()
+class LibraryTeamUserView(APIView):
+    """
+    View to add/remove/edit an individual user's permissions for a content
+    library.
+    """
+    @convert_exceptions
+    def put(self, request, lib_key_str, user_id):
+        """
+        Add a user to this content library, with permissions specified in the
+        request body.
+        """
+        key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY_TEAM)
+        serializer = ContentLibraryPermissionLevelSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        user = get_object_or_404(User, pk=int(user_id))
+        api.set_library_user_permissions(key, user, access_level=serializer.validated_data["access_level"])
+        return Response({})
+
+    @convert_exceptions
+    def delete(self, request, lib_key_str, user_id):
+        """
+        Remove the specified user's permission to access or edit this content
+        library.
+        """
+        key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY_TEAM)
+        user = get_object_or_404(User, pk=int(user_id))
+        api.set_library_user_permissions(key, user, access_level=None)
+        return Response({})
+
+
+@view_auth_classes()
+class LibraryTeamGroupView(APIView):
+    """
+    View to add/remove/edit a group's permissions for a content library.
+    """
+    @convert_exceptions
+    def put(self, request, lib_key_str, group_name):
+        """
+        Add a group to this content library, with permissions specified in the
+        request body.
+        """
+        key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY_TEAM)
+        serializer = ContentLibraryPermissionLevelSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        group = get_object_or_404(Group, name=group_name)
+        api.set_library_group_permissions(key, group, access_level=serializer.validated_data["access_level"])
+        return Response({})
+
+    @convert_exceptions
+    def delete(self, request, lib_key_str, user_id):
+        """
+        Remove the specified user's permission to access or edit this content
+        library.
+        """
+        key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY_TEAM)
+        group = get_object_or_404(Group, pk=int(user_id))
+        api.set_library_group_permissions(key, group, access_level=None)
         return Response({})
 
 
@@ -139,6 +236,7 @@ class LibraryBlockTypesView(APIView):
         Get the list of XBlock types that can be added to this library
         """
         key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
         result = api.get_allowed_block_types(key)
         return Response(LibraryXBlockTypeSerializer(result, many=True).data)
 
@@ -155,6 +253,7 @@ class LibraryCommitView(APIView):
         descendants.
         """
         key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY)
         api.publish_changes(key)
         return Response({})
 
@@ -165,6 +264,7 @@ class LibraryCommitView(APIView):
         descendants. Restore it to the last published version
         """
         key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY)
         api.revert_changes(key)
         return Response({})
 
@@ -180,6 +280,7 @@ class LibraryBlocksView(APIView):
         Get the list of all top-level blocks in this content library
         """
         key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
         result = api.get_library_blocks(key)
         return Response(LibraryXBlockMetadataSerializer(result, many=True).data)
 
@@ -189,6 +290,7 @@ class LibraryBlocksView(APIView):
         Add a new XBlock to this content library
         """
         library_key = LibraryLocatorV2.from_string(lib_key_str)
+        api.require_permission_for_library_key(library_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY)
         serializer = LibraryXBlockCreationSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         parent_block_usage_str = serializer.validated_data.pop("parent_block", None)
@@ -215,6 +317,7 @@ class LibraryBlockView(APIView):
         Get metadata about an existing XBlock in the content library
         """
         key = LibraryUsageLocatorV2.from_string(usage_key_str)
+        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
         result = api.get_library_block(key)
         return Response(LibraryXBlockMetadataSerializer(result).data)
 
@@ -232,6 +335,7 @@ class LibraryBlockView(APIView):
         be deleted but the link and the linked bundle will be unaffected.
         """
         key = LibraryUsageLocatorV2.from_string(usage_key_str)
+        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY)
         api.delete_library_block(key)
         return Response({})
 
@@ -247,6 +351,7 @@ class LibraryBlockOlxView(APIView):
         Get the block's OLX
         """
         key = LibraryUsageLocatorV2.from_string(usage_key_str)
+        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
         xml_str = api.get_library_block_olx(key)
         return Response(LibraryXBlockOlxSerializer({"olx": xml_str}).data)
 
@@ -259,6 +364,7 @@ class LibraryBlockOlxView(APIView):
         Very little validation is done.
         """
         key = LibraryUsageLocatorV2.from_string(usage_key_str)
+        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY)
         serializer = LibraryXBlockOlxSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         new_olx_str = serializer.validated_data["olx"]
@@ -280,6 +386,7 @@ class LibraryBlockAssetListView(APIView):
         List the static asset files belonging to this block.
         """
         key = LibraryUsageLocatorV2.from_string(usage_key_str)
+        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
         files = api.get_library_block_static_asset_files(key)
         return Response(LibraryXBlockStaticFilesSerializer({"files": files}).data)
 
@@ -297,6 +404,7 @@ class LibraryBlockAssetView(APIView):
         Get a static asset file belonging to this block.
         """
         key = LibraryUsageLocatorV2.from_string(usage_key_str)
+        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
         files = api.get_library_block_static_asset_files(key)
         for f in files:
             if f.path == file_path:
@@ -309,6 +417,9 @@ class LibraryBlockAssetView(APIView):
         Replace a static asset file belonging to this block.
         """
         usage_key = LibraryUsageLocatorV2.from_string(usage_key_str)
+        api.require_permission_for_library_key(
+            usage_key.lib_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY,
+        )
         file_wrapper = request.data['content']
         if file_wrapper.size > 20 * 1024 * 1024:  # > 20 MiB
             # In the future, we need a way to use file_wrapper.chunks() to read
@@ -323,11 +434,14 @@ class LibraryBlockAssetView(APIView):
         return Response(LibraryXBlockStaticFileSerializer(result).data)
 
     @convert_exceptions
-    def delete(self, request, usage_key_str, file_path):  # pylint: disable=unused-argument
+    def delete(self, request, usage_key_str, file_path):
         """
         Delete a static asset file belonging to this block.
         """
         usage_key = LibraryUsageLocatorV2.from_string(usage_key_str)
+        api.require_permission_for_library_key(
+            usage_key.lib_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY,
+        )
         try:
             api.delete_library_block_static_asset_file(usage_key, file_path)
         except ValueError:
